Check for autoviv & empty 2D arrays
In element(), row(), rows(), col(), cols(), check to see that it
doesn't autovivify the rows. In those, and in other methods like
shift and transpose, return empty object if it looks like it's going
to return no actual values; don't let it return a bunch of empty
lists.

Ensure that height and last_row give the right values for empty
arrays (0 and -1) even though there is always one empty row
even in an empty object.

Create empty() and is_empty() methods