Test dimensions with both ref and Ref::Util
Changed dimension tests so they're in a library, and added .t files
for dimensions that load either Ref::Util or just use ref, depending.

The dimension utilities don't use Ref::Util themselves, but they do
use the invocant shim, and that uses Ref::Util if available.
This serves to test that as well.